feat: consultant booking UI
