+++ frontend/src/Hooks/useToken.ts
@@ -4,7 +4,7 @@ import { createJSONStorage, persist } from 'zustand/middleware'
 
 export type TokenStore = {
 	accessToken: string | null
-	isDehydrated: boolean
+	isHydrated: boolean
 	setAccessToken: (token: string) => void
 	removeAccessToken: () => void
 }
@@ -13,21 +13,21 @@ const useToken = create<TokenStore>()(
 	persist(
 		set => ({
 			accessToken: null,
-			isDehydrated: false,
+			isHydrated: false,
 			setAccessToken: (token: string) => set({ accessToken: token }),
-			removeAccessToken: () => set({ accessToken: null, isDehydrated: false }),
+			removeAccessToken: () => set({ accessToken: null, isHydrated: false }),
 		}),
 		{
 			name: TOKEN_STORE_STRING,
 			storage: createJSONStorage(() => sessionStorage),
 			partialize: state => ({ accessToken: state.accessToken }),
 			onRehydrateStorage: () => {
-				return (_state, error) => {
+				return (state, error) => {
 					if (error) {
 						console.error('Rehydration error:', error)
 						return
 					}
-					useToken.setState({ isDehydrated: true })
+					if (state) state.isHydrated = true
 				}
 			},
 		}
