--- frontend/src/app/app/(member)/consultants/(consultants)/@consultants/default.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const DefaultPage = () => {
-	return <div>What this</div>
-}
-
-export default DefaultPage
