--- frontend/src/app/app/(member)/consultants/(consultants)/@consultants/loading.tsx
@@ -1,12 +1,0 @@
-import LoadingIcon from '@/Components/LoadingIcon'
-import React from 'react'
-
-const Loading = () => {
-	return (
-		<div className='center-all h-full w-full'>
-			<LoadingIcon />
-		</div>
-	)
-}
-
-export default Loading
