+++ frontend/src/app/app/(member)/consultants/(consultants)/layout.tsx
@@ -1,6 +1,5 @@
 'use client'
 import LoadingPage from '@/Components/Loading'
-import SearchBar from '@/Components/Management/SearchBar'
 import { useAccountStore } from '@/Hooks/useAccount'
 import {
 	isAllowedRole,
@@ -11,11 +10,9 @@ import React from 'react'
 
 const Layout = ({
 	children,
-	consultants,
 	actions,
 }: {
 	children: React.ReactNode
-	consultants: React.ReactNode
 	actions: React.ReactNode
 }) => {
 	const { isLoading, data } = useAccountStore()
@@ -27,8 +24,7 @@ const Layout = ({
 	return (
 		<div className='flex flex-col h-full w-full gap-4'>
 			<div className=''>{actions}</div>
-			<div className=''>{children}</div>
-			<div className=''>{consultants}</div>
+			<div className='flex flex-col flex-2/3'>{children}</div>
 		</div>
 	)
 }
